finished integrating new renderer with ecs changes
mat3 and mat4 now work properly as vertex attribs. now when rendering quads the origin is the center of the quad instead of lower left corner, size is specified in halfSize now. updated Animator (QuadAnimator now) to work with new way of rendering quads. using mat3 as transforms, translation/rotation/scale correctly works in solitary entities and in ones with children/parent.