audio system (CORI_81, CORI_65, CORI_66). new error handling (CORI_70)
audio system uses sdl3_mixer under the hood. new way of handling errors is designed to be used with std::expected but CoriError can be used as a regular exception as well, its derived from std::exception. PossibleErrors class allows to use std::variant with CoriError and std::expected, this allows to specify multiple possible CoriError specializations as an error in one std::expected. The logic is done, only thing left is to adopt this new error handling everywhere in the engine.