refactored assetmanager, removed selffactory, msdf-atlas caching
finished text rendering. added partial caching for msdf-atals, texture is created only once and upon further loads is loaded from cache, atlas is regenerated if the charset ranges are changed or if the original font file changed it size for some reason, original font file is determined by filename. asset manager now uses templates to have 1 set of functions for all and is using concepts for determining if an asset is valid for use with asset manager, descriptors are now defined in the asset headers themselves. removed selffactory, it made no sense basically and was limiting the flexibility of factory uses.